+++ tests/common/MockAPNSTypeChecker.swift
@@ -1,0 +1,15 @@
+//
+//  Copyright © 2022 Iterable. All rights reserved.
+//
+
+import Foundation
+
+@testable import IterableSDK
+
+struct MockAPNSTypeChecker: APNSTypeCheckerProtocol {
+    let apnsType: APNSType
+    
+    init(apnsType: APNSType) {
+        self.apnsType = apnsType
+    }
+}
