+++ tests/common/MockApplicationStateProvider.swift
@@ -1,0 +1,19 @@
+//
+//  Copyright © 2022 Iterable. All rights reserved.
+//
+
+import Foundation
+
+@testable import IterableSDK
+
+@objc public class MockApplicationStateProvider: NSObject, ApplicationStateProviderProtocol {
+    override private convenience init() {
+        self.init(applicationState: .active)
+    }
+    
+    @objc public init(applicationState: UIApplication.State) {
+        self.applicationState = applicationState
+    }
+    
+    public var applicationState: UIApplication.State
+}
