+++ tests/common/MockCustomActionDelegate.swift
@@ -1,0 +1,31 @@
+//
+//  Copyright © 2022 Iterable. All rights reserved.
+//
+
+import Foundation
+
+@testable import IterableSDK
+
+@objcMembers
+public class MockCustomActionDelegate: NSObject, IterableCustomActionDelegate {
+    // returnValue is reserved for future, don't rely on this
+    override private convenience init() {
+        self.init(returnValue: false)
+    }
+    
+    public init(returnValue: Bool) {
+        self.returnValue = returnValue
+    }
+    
+    private(set) var returnValue: Bool
+    private(set) var action: IterableAction?
+    private(set) var context: IterableActionContext?
+    var callback: ((String, IterableActionContext) -> Void)?
+    
+    public func handle(iterableCustomAction action: IterableAction, inContext context: IterableActionContext) -> Bool {
+        self.action = action
+        self.context = context
+        callback?(action.type, context)
+        return returnValue
+    }
+}
