+++ tests/common/MockInAppDelegate.swift
@@ -1,0 +1,22 @@
+//
+//  Copyright © 2022 Iterable. All rights reserved.
+//
+
+import Foundation
+
+@testable import IterableSDK
+
+class MockInAppDelegate: IterableInAppDelegate {
+    var onNewMessageCallback: ((IterableInAppMessage) -> Void)?
+    
+    init(showInApp: InAppShowResponse = .show) {
+        self.showInApp = showInApp
+    }
+    
+    func onNew(message: IterableInAppMessage) -> InAppShowResponse {
+        onNewMessageCallback?(message)
+        return showInApp
+    }
+    
+    private let showInApp: InAppShowResponse
+}
