+++ tests/common/MockInAppPersister.swift
@@ -1,0 +1,23 @@
+//
+//  Copyright © 2022 Iterable. All rights reserved.
+//
+
+import Foundation
+
+@testable import IterableSDK
+
+class MockInAppPersister: InAppPersistenceProtocol {
+    private var messages = [IterableInAppMessage]()
+    
+    func getMessages() -> [IterableInAppMessage] {
+        messages
+    }
+    
+    func persist(_ messages: [IterableInAppMessage]) {
+        self.messages = messages
+    }
+    
+    func clear() {
+        messages.removeAll()
+    }
+}
