+++ tests/common/MockNotificationResponse.swift
@@ -1,0 +1,21 @@
+//
+//  Copyright © 2022 Iterable. All rights reserved.
+//
+
+import Foundation
+
+@testable import IterableSDK
+
+struct MockNotificationResponse: NotificationResponseProtocol {
+    let userInfo: [AnyHashable: Any]
+    let actionIdentifier: String
+    
+    init(userInfo: [AnyHashable: Any], actionIdentifier: String) {
+        self.userInfo = userInfo
+        self.actionIdentifier = actionIdentifier
+    }
+    
+    var userText: String? {
+        nil
+    }
+}
