+++ tests/common/MockUrlOpener.swift
@@ -1,0 +1,22 @@
+//
+//  Copyright © 2022 Iterable. All rights reserved.
+//
+
+import Foundation
+
+@testable import IterableSDK
+
+@objcMembers
+public class MockUrlOpener: NSObject, UrlOpenerProtocol {
+    var openedUrl: URL?
+    var callback: ((URL) -> Void)?
+    
+    public init(callback: ((URL) -> Void)? = nil) {
+        self.callback = callback
+    }
+    
+    public func open(url: URL) {
+        callback?(url)
+        openedUrl = url
+    }
+}
